genesis-core/requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
PyYAML==6.0.1
httpx==0.27.0
pyaudio==0.2.13